+++ src/components/VAllResultsGrid/VImageCellSquare.vue
