+++ test/unit/specs/components/image-grid.spec.js
@@ -1,4 +1,4 @@
-import ImageGrid from '~/components/ImageGrid/ImageGrid'
+import VImageGrid from '~/components/VImageGrid/VImageGrid'
 import { render, screen } from '@testing-library/vue'
 import VueI18n from 'vue-i18n'
 import messages from '~/locales/en.json'
@@ -32,10 +32,10 @@ const options = {
     },
   },
 }
-describe('ImageGrid', () => {
+describe('VImageGrid', () => {
   it('renders images without load more button if canLoadMore is false', () => {
     options.props.canLoadMore = false
-    render(ImageGrid, options)
+    render(VImageGrid, options)
     expect(screen.queryAllByRole('img').length).toEqual(propsData.images.length)
     expect(screen.queryAllByRole('figure').length).toEqual(
       propsData.images.length
@@ -45,7 +45,7 @@ describe('ImageGrid', () => {
 
   it('renders images and load more button if canLoadMore is true', async () => {
     options.props.canLoadMore = true
-    render(ImageGrid, options)
+    render(VImageGrid, options)
     expect(screen.queryAllByRole('img').length).toEqual(propsData.images.length)
     expect(screen.queryAllByRole('figure').length).toEqual(
       propsData.images.length
